Refactor sensor and server code to their own files
